Allow cookie expiry date to be set from a float
Currently a cookie expiry date is only allowed to be an integer.
For generating a time in the future one may wish to generate a
timedelta:

datetime.timedelta(days=14).total_seconds()

This, however, produces a float:

1209600.0

To prevent the user from having to cast to an int themselves, this merge
implements the ability to set a cookie expiry date from a float, where
it is then cast into an integer to be used as usual.